feat(monorepo): added github flow to test autopublish
