+++ tools/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "scullyio/tooling",
+  "version": "1.0.0",
+  "license": "MIT",
+  "url": "",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/scullyio/scully.git"
+  },
+  "engines": {
+    "": "This is for building scully itself, using is fine with 12+",
+    "node": ">=14"
+  },
+  "type": "module"
+}
